Use standard testing bootstrap
Composer provides an `autoload-dev` configuration for the test classes, no need to mess with custom autoloading.

Also, the easiest setup for bootstrapping is to just point PHPUnit to the Composer autoloader. If people want to run the tests, they should do it from within this package. Any other scenario is pretty extravagant anyway.